+++ CHANGELOG.md
@@ -35,9 +35,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.3.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.2.3...v1.3.0) - 2017-07-09 ##
 ### Added ###
-## [Unreleased](https://github.com/craft-ai/craft-ai-client-python/compare/v1.6.0...HEAD) ##
-
-## [1.6.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.5.0...v1.6.0) - 2017-08-22 ##
+- New specialized version of the library that understands [Pandas](https://pandas.pydata.org) `DataFrame`, this specialized version of the client can be imported as such: `import Client from craftai.pandas`.
 
 ### Changed ###
 - Simplifying the import scheme, it is now possible to import the client class with `import Client from craftai`; the previous behavior still works.