+++ CHANGELOG.md
@@ -5,12 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/craft-ai/craft-ai-client-python/compare/v1.6.0...HEAD) ##
+## [Unreleased](https://github.com/craft-ai/craft-ai-client-python/compare/v1.7.0...HEAD) ##
+
+## [1.7.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.6.0...v1.7.0) - 2017-10-13 ##
 ### Added ###
-- Add new function `client.get_state_history` retrieving a agent's state history. Take a look a the [documentation](https://beta.craft.ai/doc/python#retrieve-state-history) for further informations. This function is also available in the _pandas_ version.
+## [Unreleased](https://github.com/craft-ai/craft-ai-client-python/compare/v1.7.0...HEAD) ##
+
+## [1.7.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.6.0...v1.7.0) - 2017-10-13 ##
 
 ### Fixed ###
-- Fix the interpreter error messages to fit the [test suite](https://github.com/craft-ai/craft-ai-interpreter-test-suite) expectations.
+## [Unreleased](https://github.com/craft-ai/craft-ai-client-python/compare/v1.7.0...HEAD) ##
+
+## [1.7.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.6.0...v1.7.0) - 2017-10-13 ##
 
 ## [1.6.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.5.0...v1.6.0) - 2017-08-22 ##
 ### Added ###
@@ -40,7 +46,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.3.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.2.3...v1.3.0) - 2017-07-09 ##
 ### Added ###
-- New specialized version of the library that understands [Pandas](https://pandas.pydata.org) `DataFrame`, this specialized version of the client can be imported as such: `import Client from craftai.pandas`.
+## [Unreleased](https://github.com/craft-ai/craft-ai-client-python/compare/v1.7.0...HEAD) ##
+
+## [1.7.0](https://github.com/craft-ai/craft-ai-client-python/compare/v1.6.0...v1.7.0) - 2017-10-13 ##
 
 ### Changed ###
 - Simplifying the import scheme, it is now possible to import the client class with `import Client from craftai`; the previous behavior still works.